requirements.txt
slack-bolt==1.18.0
requests==2.31.0
python-dotenv==1.0.1
schedule==1.2.1
